PersonalizedTraining/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
google-generativeai==0.3.2
httpx==0.25.1